requirements.txt
telethon
pymongo
dnspython
requests
python-dotenv